// src/LawyerSite.Web/src/app/services/client/web-api-client.ts
export interface ServiceDto {
    id?: number;
    title?: string;
    description?: string | undefined;
    price?: number;
    isActive?: boolean;
    durationMinutes?: number;
    createdAt?: Date;
    updatedAt?: Date;
}